guide works now yay

+++ src/main/java/ca/mcgill/ecse/snowshoetours/controller/GearController.java
@@ -122,7 +122,7 @@ public class GearController {
       if(combo.getName().equals(name)) aCombo = combo;
     }
     
-    aCombo.delete();
+    if (aCombo != null) aCombo.delete();
   }
 
   // this method does not need to be implemented by a team with five team members
